+++ src/rpc/responses.rs
@@ -47,7 +47,7 @@ use crate::{
     pubkey::Pubkey,
     py_from_bytes_general_via_bincode, pybytes_general_via_bincode,
     signature::Signature,
-    tmp_account_decoder::{UiAccount, UiTokenAmount as UiTokenAmountOriginal},
+    tmp_account_decoder::{UiAccount, UiAccountData, UiTokenAmount as UiTokenAmountOriginal},
     tmp_transaction_status::{
         TransactionConfirmationStatus as TransactionConfirmationStatusOriginal,
         TransactionStatus as TransactionStatusOriginal, UiTransactionReturnData,
@@ -285,7 +285,7 @@ response_data_boilerplate!(RpcResponseContext);
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
 #[serde(untagged)]
-pub enum Resp<T: PyClass + IntoPy<PyObject>> {
+pub enum Resp<T: IntoPy<PyObject>> {
     Result {
         #[serde(skip_deserializing)]
         jsonrpc: crate::rpc::requests::V2,
@@ -641,7 +641,7 @@ macro_rules! notification_contextless {
 contextful_resp_eq!(
     GetAccountInfoResp,
     Option<Account>,
-    "Option<FromInto<UiAccount>>"
+    "Option<TryFromInto<UiAccount>>"
 );
 
 contextful_resp_eq!(
@@ -650,48 +650,10 @@ contextful_resp_eq!(
     "Option<TryFromInto<UiAccount>>"
 );
 
-macro_rules! parse_maybe_json {
-    ($resp:ident, $func:ident) => {
-        paste! {
-            #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
-            #[serde(untagged)]
-            pub enum [<$resp MaybeJsonParsedResp>] {
-                Parsed(Resp<[<$resp JsonParsedResp>]>),
-                Binary(Resp<[<$resp Resp>]>),
-            }
-
-            impl IntoPy<PyObject> for [<$resp MaybeJsonParsedResp>] {
-                fn into_py(self, py: Python<'_>) -> PyObject {
-                    match self {
-                        Self::Parsed(x) => x.into_py(py),
-                        Self::Binary(x) => x.into_py(py),
-                    }
-                }
-            }
-
-            #[doc = "Parse a ``" $resp "`` response that may or may not be in jsonParsed format.
-Args:
-    raw (str): The raw JSON.
-
-Returns:
-    Union[" $resp "Resp, " $resp "JsonParsedResp]: The parsed result.          
-"]
-            #[pyfunction]
-            pub fn [<parse_ $func _maybe_json>](raw: &str) -> PyResult<[<$resp MaybeJsonParsedResp>]> {
-                serde_json::from_str(raw).map_err(to_py_err)
-            }
-        }
-    };
-}
-
-parse_maybe_json!(GetAccountInfo, account_info);
-parse_maybe_json!(GetMultipleAccounts, multiple_accounts);
-parse_maybe_json!(GetTokenAccountsByDelegate, token_accounts_by_delegate);
-parse_maybe_json!(GetTokenAccountsByOwner, token_accounts_by_owner);
-parse_maybe_json!(GetProgramAccountsWithContext, program_accounts_with_context);
-parse_maybe_json!(
-    GetProgramAccountsWithoutContext,
-    program_accounts_without_context
+contextful_resp_eq!(
+    GetAccountInfoMaybeJsonParsedResp,
+    Option<AccountMaybeJSON>,
+    "Option<TryFromInto<UiAccount>>"
 );
 
 contextful_resp_eq!(GetBalanceResp, u64);
@@ -1015,7 +977,7 @@ pub struct RpcSimulateTransactionResult {
     pub err: Option<TransactionErrorType>,
     #[pyo3(get)]
     pub logs: Option<Vec<String>>,
-    #[serde_as(as = "Option<Vec<Option<FromInto<UiAccount>>>>")]
+    #[serde_as(as = "Option<Vec<Option<TryFromInto<UiAccount>>>>")]
     #[pyo3(get)]
     pub accounts: Option<Vec<Option<Account>>>,
     #[pyo3(get)]
@@ -1324,11 +1286,86 @@ contextless_resp_eq!(GetMinimumBalanceForRentExemptionResp, u64);
 contextful_resp_eq!(
     GetMultipleAccountsResp,
     Vec<Option<Account>>,
-    "Vec<Option<FromInto<UiAccount>>>"
+    "Vec<Option<TryFromInto<UiAccount>>>"
 );
 contextful_resp_eq!(
     GetMultipleAccountsJsonParsedResp,
     Vec<Option<AccountJSON>>,
+    "Vec<Option<TryFromInto<UiAccount>>>"
+);
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, FromPyObject)]
+#[serde(untagged)]
+pub enum AccountMaybeJSON {
+    Binary(Account),
+    Parsed(AccountJSON),
+}
+
+impl IntoPy<PyObject> for AccountMaybeJSON {
+    fn into_py(self, py: Python<'_>) -> PyObject {
+        match self {
+            Self::Parsed(x) => x.into_py(py),
+            Self::Binary(x) => x.into_py(py),
+        }
+    }
+}
+
+impl From<Account> for AccountMaybeJSON {
+    fn from(a: Account) -> Self {
+        Self::Binary(a)
+    }
+}
+
+impl From<AccountJSON> for AccountMaybeJSON {
+    fn from(a: AccountJSON) -> Self {
+        Self::Parsed(a)
+    }
+}
+
+impl TryFrom<AccountMaybeJSON> for Account {
+    type Error = String;
+    fn try_from(acc: AccountMaybeJSON) -> Result<Self, Self::Error> {
+        if let AccountMaybeJSON::Binary(account) = acc {
+            Ok(account)
+        } else {
+            Err("Expected Account, found AccountJSON".to_string())
+        }
+    }
+}
+
+impl TryFrom<AccountMaybeJSON> for AccountJSON {
+    type Error = String;
+    fn try_from(acc: AccountMaybeJSON) -> Result<Self, Self::Error> {
+        if let AccountMaybeJSON::Parsed(account) = acc {
+            Ok(account)
+        } else {
+            Err("Expected AccountJSON, found Account".to_string())
+        }
+    }
+}
+
+impl From<UiAccount> for AccountMaybeJSON {
+    fn from(u: UiAccount) -> Self {
+        match u.data {
+            UiAccountData::LegacyBinary(_) => panic!("LegacyBinary data should not appear"),
+            UiAccountData::Json(_) => AccountJSON::try_from(u).unwrap().into(),
+            UiAccountData::Binary(..) => Account::try_from(u).unwrap().into(),
+        }
+    }
+}
+
+impl From<AccountMaybeJSON> for UiAccount {
+    fn from(a: AccountMaybeJSON) -> Self {
+        match a {
+            AccountMaybeJSON::Binary(acc) => Self::from(acc),
+            AccountMaybeJSON::Parsed(acc) => Self::try_from(acc).unwrap(),
+        }
+    }
+}
+
+contextful_resp_eq!(
+    GetMultipleAccountsMaybeJsonParsedResp,
+    Vec<Option<AccountMaybeJSON>>,
     "Vec<Option<TryFromInto<UiAccount>>>"
 );
 
@@ -1341,7 +1378,7 @@ pub struct RpcKeyedAccount {
     #[serde_as(as = "DisplayFromStr")]
     #[pyo3(get)]
     pub pubkey: Pubkey,
-    #[serde_as(as = "FromInto<UiAccount>")]
+    #[serde_as(as = "TryFromInto<UiAccount>")]
     #[pyo3(get)]
     pub account: Account,
 }
@@ -1389,14 +1426,36 @@ contextful_resp_eq!(
     Vec<RpcKeyedAccountJsonParsed>
 );
 
+contextless_resp_eq!(GetProgramAccountsResp, Vec<RpcKeyedAccount>, clone);
 contextless_resp_eq!(
-    GetProgramAccountsWithoutContextResp,
-    Vec<RpcKeyedAccount>,
+    GetProgramAccountsJsonParsedResp,
+    Vec<RpcKeyedAccountJsonParsed>,
     clone
 );
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, FromPyObject)]
+#[serde(untagged)]
+pub enum RpcKeyedAccountMaybeJSON {
+    Binary(RpcKeyedAccount),
+    Parsed(RpcKeyedAccountJsonParsed),
+}
+
+impl IntoPy<PyObject> for RpcKeyedAccountMaybeJSON {
+    fn into_py(self, py: Python<'_>) -> PyObject {
+        match self {
+            Self::Parsed(x) => x.into_py(py),
+            Self::Binary(x) => x.into_py(py),
+        }
+    }
+}
+
+contextful_resp_eq!(
+    GetProgramAccountsWithContextMaybeJsonParsedResp,
+    Vec<RpcKeyedAccountMaybeJSON>
+);
 contextless_resp_eq!(
-    GetProgramAccountsWithoutContextJsonParsedResp,
-    Vec<RpcKeyedAccountJsonParsed>,
+    GetProgramAccountsMaybeJsonParsedResp,
+    Vec<RpcKeyedAccountMaybeJSON>,
     clone
 );
 
@@ -2379,7 +2438,7 @@ contextless_resp_eq!(RequestAirdropResp, Signature, "DisplayFromStr");
 contextless_resp_eq!(SendTransactionResp, Signature, "DisplayFromStr");
 contextful_resp_eq!(SimulateTransactionResp, RpcSimulateTransactionResult);
 
-notification!(AccountNotification, Account, "FromInto<UiAccount>");
+notification!(AccountNotification, Account, "TryFromInto<UiAccount>");
 notification!(
     AccountNotificationJsonParsed,
     AccountJSON,
@@ -2430,17 +2489,27 @@ impl IntoPy<PyObject> for ProgramNotificationType {
 contextless_resp_eq!(SubscriptionResp, u64);
 
 macro_rules ! pyunion_resp {
-    ($name:ident, $($variant:ident),+) => {
+    ($name:ident, $err_variant:ident, $($variant:ident),+) => {
         #[derive(FromPyObject, Clone, Debug, PartialEq, Serialize, Deserialize)]
         #[serde(untagged, rename_all = "camelCase")]
         pub enum $name {
+            $err_variant($err_variant),
             $($variant($variant),)+
         }
 
         impl $name {
             fn to_json(&self) -> String {
                 match self {
+                    Self::$err_variant(x) => x.py_to_json(),
                     $(Self::$variant(x) => x.py_to_json(),)+
+                }
+            }
+
+            fn from_json(raw: &str, parser: &str) -> PyResult<Self> {
+                match parser {
+                    stringify!($err_variant) => {let parsed = $err_variant::py_from_json(raw)?; let as_enum = Self::RpcError(parsed); Ok(as_enum)},
+                    $(stringify!($variant) => {let parsed = $variant::py_from_json(raw)?; let as_enum = match parsed {Resp::Error {error, ..} => Self::RpcError(error), Resp::Result {result, ..} => Self::$variant(result)};Ok(as_enum)},)+
+                    _ => Err(PyValueError::new_err(format!("Unrecognised parser: {}", parser)))
                 }
             }
         }
@@ -2448,6 +2517,7 @@ macro_rules ! pyunion_resp {
         impl IntoPy<PyObject> for $name {
             fn into_py(self, py: Python<'_>) -> PyObject {
                 match self {
+                    Self::$err_variant(x) => x.into_py(py),
                     $(Self::$variant(x) => x.into_py(py),)+
                 }
             }
@@ -2460,6 +2530,7 @@ pyunion_resp!(
     RpcError,
     GetAccountInfoResp,
     GetAccountInfoJsonParsedResp,
+    GetAccountInfoMaybeJsonParsedResp,
     GetBalanceResp,
     GetBlockProductionResp,
     GetBlockResp,
@@ -2488,10 +2559,13 @@ pyunion_resp!(
     GetMinimumBalanceForRentExemptionResp,
     GetMultipleAccountsResp,
     GetMultipleAccountsJsonParsedResp,
+    GetMultipleAccountsMaybeJsonParsedResp,
     GetProgramAccountsWithContextResp,
-    GetProgramAccountsWithoutContextResp,
+    GetProgramAccountsResp,
     GetProgramAccountsWithContextJsonParsedResp,
-    GetProgramAccountsWithoutContextJsonParsedResp,
+    GetProgramAccountsJsonParsedResp,
+    GetProgramAccountsWithContextMaybeJsonParsedResp,
+    GetProgramAccountsMaybeJsonParsedResp,
     GetRecentPerformanceSamplesResp,
     GetSignaturesForAddressResp,
     GetSignatureStatusesResp,
@@ -2510,7 +2584,6 @@ pyunion_resp!(
     GetTransactionResp,
     GetTransactionCountResp,
     GetVersionResp,
-    RpcVersionInfo,
     GetVoteAccountsResp,
     IsBlockhashValidResp,
     MinimumLedgerSlotResp,
@@ -2546,9 +2619,10 @@ pub fn batch_to_json(resps: Vec<RPCResult>) -> String {
 ///
 /// Args:
 ///     raw (str): The batch JSON string.
+///     parsers (Sequence): The classes to parse.
 ///
 /// Returns:
-///     A list of request objects.
+///     A list of response objects.
 ///
 /// Example:
 ///     >>> from solders.rpc.responses import batch_from_json, GetBlockHeightResp, GetFirstAvailableBlockResp
@@ -2570,18 +2644,10 @@ pub fn batch_from_json(raw: &str, parsers: Vec<&PyType>) -> PyResult<Vec<PyObjec
         let msg = format!("Number of parsers does not match number of response objects. Num parsers: {}. Num responses: {}", parsers_len, raw_objects_len);
         Err(PyValueError::new_err(msg))
     } else {
-        Python::with_gil(|py| {
-            Ok(raw_objects
-                .iter()
-                .zip(parsers.iter())
-                .map(|(res, parser)| {
-                    parser
-                        .call_method1("from_json", (serde_json::to_string(res).unwrap(),))
-                        .unwrap()
-                        .into_py(py)
-                })
-                .collect())
-        })
+        let parsed = raw_objects.iter().zip(parsers.iter()).map(|(res, parser)| {
+            RPCResult::from_json(&serde_json::to_string(res).unwrap(), parser.name().unwrap())
+        });
+        Python::with_gil(|py| parsed.map(|obj| obj.map(|o| o.into_py(py))).collect())
     }
 }
 
@@ -2649,6 +2715,7 @@ pub(crate) fn create_responses_mod(py: Python<'_>) -> PyResult<&PyModule> {
             RpcError::type_object(py),
             GetAccountInfoResp::type_object(py),
             GetAccountInfoJsonParsedResp::type_object(py),
+            GetAccountInfoMaybeJsonParsedResp::type_object(py),
             GetBalanceResp::type_object(py),
             GetBlockProductionResp::type_object(py),
             GetBlockResp::type_object(py),
@@ -2677,10 +2744,13 @@ pub(crate) fn create_responses_mod(py: Python<'_>) -> PyResult<&PyModule> {
             GetMinimumBalanceForRentExemptionResp::type_object(py),
             GetMultipleAccountsResp::type_object(py),
             GetMultipleAccountsJsonParsedResp::type_object(py),
+            GetMultipleAccountsMaybeJsonParsedResp::type_object(py),
             GetProgramAccountsWithContextResp::type_object(py),
-            GetProgramAccountsWithoutContextResp::type_object(py),
+            GetProgramAccountsResp::type_object(py),
             GetProgramAccountsWithContextJsonParsedResp::type_object(py),
-            GetProgramAccountsWithoutContextJsonParsedResp::type_object(py),
+            GetProgramAccountsJsonParsedResp::type_object(py),
+            GetProgramAccountsMaybeJsonParsedResp::type_object(py),
+            GetProgramAccountsWithContextMaybeJsonParsedResp::type_object(py),
             GetRecentPerformanceSamplesResp::type_object(py),
             GetSignaturesForAddressResp::type_object(py),
             GetSignatureStatusesResp::type_object(py),
@@ -2756,6 +2826,7 @@ pub(crate) fn create_responses_mod(py: Python<'_>) -> PyResult<&PyModule> {
     m.add_class::<RpcError>()?;
     m.add_class::<GetAccountInfoResp>()?;
     m.add_class::<GetAccountInfoJsonParsedResp>()?;
+    m.add_class::<GetAccountInfoMaybeJsonParsedResp>()?;
     m.add_class::<GetBalanceResp>()?;
     m.add_class::<RpcBlockProduction>()?;
     m.add_class::<RpcBlockProductionRange>()?;
@@ -2796,12 +2867,15 @@ pub(crate) fn create_responses_mod(py: Python<'_>) -> PyResult<&PyModule> {
     m.add_class::<GetMinimumBalanceForRentExemptionResp>()?;
     m.add_class::<GetMultipleAccountsResp>()?;
     m.add_class::<GetMultipleAccountsJsonParsedResp>()?;
+    m.add_class::<GetMultipleAccountsMaybeJsonParsedResp>()?;
     m.add_class::<RpcKeyedAccount>()?;
     m.add_class::<RpcKeyedAccountJsonParsed>()?;
     m.add_class::<GetProgramAccountsWithContextResp>()?;
-    m.add_class::<GetProgramAccountsWithoutContextResp>()?;
+    m.add_class::<GetProgramAccountsResp>()?;
     m.add_class::<GetProgramAccountsWithContextJsonParsedResp>()?;
-    m.add_class::<GetProgramAccountsWithoutContextJsonParsedResp>()?;
+    m.add_class::<GetProgramAccountsJsonParsedResp>()?;
+    m.add_class::<GetProgramAccountsWithContextMaybeJsonParsedResp>()?;
+    m.add_class::<GetProgramAccountsMaybeJsonParsedResp>()?;
     m.add_class::<RpcPerfSample>()?;
     m.add_class::<GetRecentPerformanceSamplesResp>()?;
     m.add_class::<RpcConfirmedTransactionStatusWithSignature>()?;
@@ -2879,13 +2953,6 @@ pub(crate) fn create_responses_mod(py: Python<'_>) -> PyResult<&PyModule> {
         wrap_pyfunction!(batch_from_json, m)?,
         wrap_pyfunction!(parse_websocket_message, m)?,
         wrap_pyfunction!(parse_notification, m)?,
-        wrap_pyfunction!(parse_account_info_maybe_json, m)?,
-        wrap_pyfunction!(parse_multiple_accounts_maybe_json, m)?,
-        wrap_pyfunction!(parse_token_accounts_by_delegate_maybe_json, m)?,
-        wrap_pyfunction!(parse_token_accounts_by_owner_maybe_json, m)?,
-        wrap_pyfunction!(parse_account_info_maybe_json, m)?,
-        wrap_pyfunction!(parse_program_accounts_with_context_maybe_json, m)?,
-        wrap_pyfunction!(parse_program_accounts_without_context_maybe_json, m)?,
     ];
     for func in funcs {
         m.add_function(func)?;