+++ tests/test_rpc_responses.py
@@ -5,6 +5,7 @@ from solders.errors import SerdeJSONError
 from solders.rpc.responses import (
     GetAccountInfoResp,
     GetAccountInfoJsonParsedResp,
+    GetAccountInfoMaybeJsonParsedResp,
     GetBalanceResp,
     GetBlockResp,
     GetBlockCommitmentResp,
@@ -35,10 +36,13 @@ from solders.rpc.responses import (
     GetMinimumBalanceForRentExemptionResp,
     GetMultipleAccountsResp,
     GetMultipleAccountsJsonParsedResp,
+    GetMultipleAccountsMaybeJsonParsedResp,
     GetProgramAccountsWithContextResp,
-    GetProgramAccountsWithoutContextResp,
+    GetProgramAccountsResp,
     GetProgramAccountsWithContextJsonParsedResp,
-    GetProgramAccountsWithoutContextJsonParsedResp,
+    GetProgramAccountsJsonParsedResp,
+    GetProgramAccountsWithContextMaybeJsonParsedResp,
+    GetProgramAccountsMaybeJsonParsedResp,
     GetRecentPerformanceSamplesResp,
     GetSignaturesForAddressResp,
     GetSignatureStatusesResp,
@@ -114,12 +118,6 @@ from solders.rpc.responses import (
     batch_to_json,
     parse_notification,
     parse_websocket_message,
-    parse_account_info_maybe_json,
-    parse_multiple_accounts_maybe_json,
-    parse_token_accounts_by_delegate_maybe_json,
-    parse_token_accounts_by_owner_maybe_json,
-    parse_program_accounts_with_context_maybe_json,
-    parse_program_accounts_without_context_maybe_json,
 )
 from solders.rpc.errors import NodeUnhealthy
 from solders.hash import Hash
@@ -190,8 +188,9 @@ def test_get_account_info() -> None:
     # this happens when jsonParsed is requested but the RPC can't do it.
     with raises(SerdeJSONError):
         GetAccountInfoJsonParsedResp.from_json(raw)
-    parsed2 = parse_account_info_maybe_json(raw)
-    assert parsed == parsed2
+    parsed2 = GetAccountInfoMaybeJsonParsedResp.from_json(raw)
+    assert isinstance(parsed2, GetAccountInfoMaybeJsonParsedResp)
+    assert parsed.value == parsed2.value
 
 
 def test_get_account_info_null() -> None:
@@ -234,8 +233,9 @@ def test_get_account_info_json_parsed() -> None:
     assert parsed == GetAccountInfoJsonParsedResp(context=context, value=account_json)
     assert parsed.context.slot == 140702417
     assert parsed.value.data.program == "spl-token"
-    parsed2 = parse_account_info_maybe_json(raw)
-    assert parsed == parsed2
+    parsed2 = GetAccountInfoMaybeJsonParsedResp.from_json(raw)
+    assert isinstance(parsed2, GetAccountInfoMaybeJsonParsedResp)
+    assert parsed.value == parsed2.value
 
 
 def test_get_balance() -> None:
@@ -1000,8 +1000,19 @@ def test_get_multiple_accounts_base64() -> None:
         rent_epoch=2,
         data=b"",
     )
-    parsed2 = parse_multiple_accounts_maybe_json(raw)
-    assert parsed == parsed2
+    parsed2 = GetMultipleAccountsMaybeJsonParsedResp.from_json(raw)
+    assert isinstance(parsed2, GetMultipleAccountsMaybeJsonParsedResp)
+    assert parsed.value == parsed2.value
+
+
+def test_get_multiple_accounts_maybe_json() -> None:
+    # mix of jsonParsed and base64
+    raw = '{"jsonrpc":"2.0","result":{"context":{"apiVersion":"1.13.2","slot":167617386},"value":[{"data":{"parsed":{"info":{"decimals":9,"freezeAuthority":null,"isInitialized":true,"mintAuthority":null,"supply":"0"},"type":"mint"},"program":"spl-token","space":82},"executable":false,"lamports":736771963562,"owner":"TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA","rentEpoch":634},{"data":["AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAJAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA==","base64"],"executable":false,"lamports":998763433,"owner":"2WRuhE4GJFoE23DYzp2ij6ZnuQ8p9mJeU6gDgfsjR4or","rentEpoch":371}]},"id":1}'
+    parsed = GetMultipleAccountsMaybeJsonParsedResp.from_json(raw)
+    assert isinstance(parsed, GetMultipleAccountsMaybeJsonParsedResp)
+    val = parsed.value
+    assert isinstance(val[0], AccountJSON)
+    assert isinstance(val[1], Account)
 
 
 def test_get_multiple_accounts_base58() -> None:
@@ -1052,8 +1063,9 @@ def test_get_multiple_accounts_base58() -> None:
         rent_epoch=2,
         data=b"",
     )
-    parsed2 = parse_multiple_accounts_maybe_json(raw)
-    assert parsed == parsed2
+    parsed2 = GetMultipleAccountsMaybeJsonParsedResp.from_json(raw)
+    assert isinstance(parsed2, GetMultipleAccountsMaybeJsonParsedResp)
+    assert parsed.value == parsed2.value
 
 
 def test_get_multiple_accounts_json_parsed() -> None:
@@ -1110,8 +1122,9 @@ def test_get_multiple_accounts_json_parsed() -> None:
     assert data.program == "spl-token"
     assert data.space == 165
     assert isinstance(data.parsed, str)
-    parsed2 = parse_multiple_accounts_maybe_json(raw)
-    assert parsed == parsed2
+    parsed2 = GetMultipleAccountsMaybeJsonParsedResp.from_json(raw)
+    assert isinstance(parsed2, GetMultipleAccountsMaybeJsonParsedResp)
+    assert parsed.value == parsed2.value
 
 
 def test_get_program_accounts_without_context() -> None:
@@ -1131,8 +1144,8 @@ def test_get_program_accounts_without_context() -> None:
   ],
   "id": 1
 }"""
-    parsed = GetProgramAccountsWithoutContextResp.from_json(raw)
-    assert isinstance(parsed, GetProgramAccountsWithoutContextResp)
+    parsed = GetProgramAccountsResp.from_json(raw)
+    assert isinstance(parsed, GetProgramAccountsResp)
     assert parsed.value[0] == RpcKeyedAccount(
         account=Account(
             data=b58decode(b"2R9jLfiAQ9bgdcw6h8s44439"),
@@ -1190,8 +1203,8 @@ def test_get_program_accounts_without_context_json_parsed() -> None:
         }
     ]
 }"""
-    parsed = GetProgramAccountsWithoutContextJsonParsedResp.from_json(raw)
-    assert isinstance(parsed, GetProgramAccountsWithoutContextJsonParsedResp)
+    parsed = GetProgramAccountsJsonParsedResp.from_json(raw)
+    assert isinstance(parsed, GetProgramAccountsJsonParsedResp)
     val = parsed.value[0]
     assert isinstance(val, RpcKeyedAccountJsonParsed)
     acc = val.account
@@ -1207,8 +1220,9 @@ def test_get_program_accounts_without_context_json_parsed() -> None:
     assert data.program == "spl-token-2022"
     assert data.space == 182
     assert isinstance(data.parsed, str)
-    parsed2 = parse_program_accounts_without_context_maybe_json(raw)
-    assert parsed == parsed2
+    parsed2 = GetProgramAccountsMaybeJsonParsedResp.from_json(raw)
+    assert isinstance(parsed2, GetProgramAccountsMaybeJsonParsedResp)
+    assert parsed.value == parsed2.value
 
 
 def test_get_program_accounts_with_context() -> None:
@@ -1216,8 +1230,8 @@ def test_get_program_accounts_with_context() -> None:
     parsed = GetProgramAccountsWithContextResp.from_json(raw)
     assert isinstance(parsed, GetProgramAccountsWithContextResp)
     assert not parsed.value
-    parsed2 = parse_program_accounts_with_context_maybe_json(raw)
-    # note: parsed2 thinks it's JsonParsed but doesn't matter because value is empty
+    parsed2 = GetProgramAccountsWithContextMaybeJsonParsedResp.from_json(raw)
+    assert isinstance(parsed2, GetProgramAccountsWithContextMaybeJsonParsedResp)
     assert parsed.context == parsed2.context and parsed.value == parsed2.value
 
 
@@ -1289,8 +1303,9 @@ def test_get_program_accounts_with_context_json_parsed() -> None:
     assert data.program == "spl-token-2022"
     assert data.space == 182
     assert isinstance(data.parsed, str)
-    parsed2 = parse_program_accounts_with_context_maybe_json(raw)
-    assert parsed == parsed2
+    parsed2 = GetProgramAccountsWithContextMaybeJsonParsedResp.from_json(raw)
+    assert isinstance(parsed2, GetProgramAccountsWithContextMaybeJsonParsedResp)
+    assert parsed.value == parsed2.value
 
 
 def test_get_recent_performance_samples() -> None:
@@ -1588,8 +1603,6 @@ def test_get_token_accounts_by_owner_json_parsed() -> None:
     assert acc.owner == Pubkey.from_string(
         "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
     )
-    parsed2 = parse_token_accounts_by_owner_maybe_json(raw)
-    assert parsed == parsed2
 
 
 def test_get_token_accounts_by_owner_base64() -> None:
@@ -1638,8 +1651,6 @@ def test_get_token_accounts_by_owner_base64() -> None:
         executable=False,
         rent_epoch=341,
     )
-    parsed2 = parse_token_accounts_by_owner_maybe_json(raw)
-    assert parsed == parsed2
 
 
 def test_get_token_accounts_by_delegate_json_parsed() -> None:
@@ -1708,8 +1719,6 @@ def test_get_token_accounts_by_delegate_json_parsed() -> None:
     assert data.program == "spl-token"
     assert isinstance(data.parsed, str)
     assert data.space == 165
-    parsed2 = parse_token_accounts_by_delegate_maybe_json(raw)
-    assert parsed == parsed2
 
 
 def test_get_token_accounts_by_delegate_base64() -> None:
@@ -1758,8 +1767,6 @@ def test_get_token_accounts_by_delegate_base64() -> None:
         executable=False,
         rent_epoch=341,
     )
-    parsed2 = parse_token_accounts_by_delegate_maybe_json(raw)
-    assert parsed == parsed2
 
 
 def test_get_token_largest_accounts() -> None: